v0.06
The player can now interact with an instance, and by selecting the "look" option, they are provided a simple scene where they can interact with objects. (ignore the fact that its l ron hubbard right now that was the first image i found to use) upon taking an object, that object will no longer appear in the scene, even after reloading. this feature is highly likely to be changed in the future, as right now, it requires the scene to become invisible rather than be fully destroyed. But for what its worth I think it works great. Also added some debug stuff for instance depth sorting, im planning on expanding the debug feature further when I have reason to do so.