v0.1
Implemented interactive cutscenes, specifically for the news broadcasts the player will be in charge of managing throughout the game. It is very rudimentary, but this proof of concept is enough for now. Also made a simple layout of the newsroom scene with a custom background, but this will need a lot of polishing.