v0.11
Minor changes to the textbox element. The speech bubble will now appear at the bottom, only when the dialogue is related to the game. If not, it will appear next to the character that is speaking; however, this resulted in me having to create a new text event every time a new character begins talking. May change this in the future.